// src/interfaces/IPatient.ts
export default interface IPatient {
  id?: number;
  name: string;
  age: number;
  cpf: string;
  rg: string;
  birthDate: string;
  sex: string;
  sign: string;
  motherName: string;
  fatherName: string;
  email: string;
  password: string;
  zipCode: string;
  address: string;
  addressNumber: number;
  district: string;
  city: string;
  state: string;
  phone: string;
  mobilePhone: string;
  height: string;
  weight: string;
  bloodType: string;
  color: string;
}